fastapi_project/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
asyncpg==0.31.0
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
executing==2.2.1
fastapi==0.128.1
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
icecream==2.1.10
idna==3.11
iniconfig==2.3.0
orjson==3.11.7
packaging==26.0
pluggy==1.6.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
SQLAlchemy==2.0.46
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uv==0.10.9
uvicorn==0.40.0
